pyproject.toml
[tool.poetry]
name = "cf-pdf-demo"
version = "0.1.0"
description = ""
authors = ["Danielle Hon"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pinecone = "^5.1.0"
openai = "^1.45.0"
requests = "^2.32.3"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
langchain-pinecone = "^0.2.0"
langchain-huggingface = "^0.1.0"
langchain-text-splitters = "^0.3.0"
groq = "^0.11.0"
tiktoken = "^0.7.0"
google-generativeai = "^0.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
